requirements.txt
Pillow
Requests
streamlit
streamlit_lottie
streamlit_timeline
nltk
pytest
pdfreader
